Move Callable-classes to helper-class to cleanup I2CDeviceImpl
